requirements.txt
httpx>=0.25.0
python-dotenv>=1.0.0
langchain-core>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.20
langchain-community>=0.0.20
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
pypdf>=3.0.0
beautifulsoup4>=4.12.0
tavily-python>=0.3.0
langchain-tavily>=0.1.0
streamlit>=1.28.0
streamlit-shadcn-ui>=0.1.0
streamlit-option-menu>=0.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
responses>=0.23.0
